move to puppeteer

+++ test/environment.js
@@ -1,5 +1,6 @@
 import server from '../src/server';
 import {spawn} from 'child_process';
+import {sleep} from 'sleep';
 
 const app = server();
 export const beforeAndAfter = (port = 3100) => {
@@ -10,11 +11,14 @@ export const beforeAndAfter = (port = 3100) => {
 
   after(() => expressServer.close());
 
-  before(() => {
-    cdnServer = spawn('node', ['./dist/test/browser/mock/cdn-server.js'], {stdio: 'inherit'});
-  });
+  if (process.env.WIX_NODE_BUILD_WATCH_MODE !== 'true') {
+    before(() => {
+      cdnServer = spawn('node', ['./dist/test/browser/mock/cdn-server.js'], {stdio: 'inherit'});
+      sleep(1);
+    });
 
-  after(() => {
-    cdnServer.kill();
-  });
+    after(() => {
+      cdnServer.kill();
+    });
+  }
 };
